Make DDC config available on new ArtooDaq instance
Call to ArtooDaq.read_ddc_1st_config will now work even if no prior
call to ArtooDaq.tune_ddc_1st_to_freq occurred on that instance. In
this case the ideal settings fields in the returned config dictionary
are not readily available, but the actual (or 'digital') settings
that correspond to configured software registers in the ROACH2 are
populated.

The channel frequency (mid-point of the Nyquist zone where the
signal-of-interest resides) is estimated by 1) computing the frequency
response from the digital filter coefficients, 2) finding the
frequency where this response has the maximum magnitude, 3) and
rounding that point to the center of the nearest 100MHz channel (e.g.
150MHz, 250MHz, ..., 1450e6).

A new field ['digital']['f_c'] has also been added to the returned
configuration dictionary which indicates the true center of the
1st stage DDC, accounting for the DDS frequency resolution.