Inherit from EthernetSCPIService
I modified the EthernetThermoFisherService to now inherit from
EthernetSCPIService. This now uses the _reconnect method and no changes
to send_to_device are needed. I make use of the
EthernetSCPIService.__init__ . Also the acknowledge at reconnect is now
tested. _send_commands is overwritten but simplified compared to earlier
version, since the connection is now handled elsewere. The _listen
method is not used.